Modify the logic of deleteSeller and add jwt authentication

+++ routes/seller.js
@@ -169,13 +169,27 @@ router.updateSeller = (req, res) => {
 router.deleteSeller = (req, res) => {
     res.setHeader('Content-Type', 'application/json')
 
-    Seller.findByIdAndRemove(req.params.id, (err, seller) => {
-        if (err) {
-            res.send(JSON.stringify({code: ERR_NOK, error: err}, null, 5))
-        } else {
-            res.send(JSON.stringify({code: ERR_OK, message: "Successfully Delete Seller"}, null, 5))
-        }
-    })
+    // jwt
+    let token = req.body.token
+    if (!token) {
+        res.send(JSON.stringify({code: USER_NAT, message: 'Not Login Yet, Please Login'}, null, 5))
+    } else {
+        jwt.verify(token, config.superSecret, (err, decoded) => {
+            if (err) {
+                res.send(JSON.stringify({code: ERR_NOK, error: err}, null, 5))
+            } else {
+                req.decoded = decoded
+
+                Seller.findByIdAndRemove(req.params.id, (err, seller) => {
+                    if (err) {
+                        res.send(JSON.stringify({code: ERR_NOK, error: err}, null, 5))
+                    } else {
+                        res.send(JSON.stringify({code: ERR_OK, message: "Successfully Delete Seller"}, null, 5))
+                    }
+                })
+            }
+        })
+    }
 }
 
 /**
